% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F2.R
\name{F2}
\alias{F2}
\title{Función de ejmplo: operación aritmética RESTA}
\usage{
F2(x_1, x_2)
}
\arguments{
\item{x_1}{Primer valor}

\item{x_2}{Segundo valor}
}
\value{
Retorna el resultado de la resta
}
\description{
Función de ejmplo: operación aritmética RESTA
}
\examples{
F_example(x_1 = 3, x_2 = 4)
F_example(x_1 = 1, x_2 = 2)
}
