% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_example.R
\name{F_example}
\alias{F_example}
\title{Función de ejmplo: operación aritmética SUMA}
\usage{
F_example(x_1, x_2)
}
\arguments{
\item{x_1}{Primer valor}

\item{x_2}{Segundo valor}
}
\value{
Retorna el resultado de la suma
}
\description{
Función de ejmplo: operación aritmética SUMA
}
\examples{
F_example(x_1 = 3, x_2 = 4)
F_example(x_1 = 1, x_2 = 2)
}
